+++ controllers/nifinodegroupautoscaler_controller.go
@@ -86,7 +86,7 @@ func (r *NifiNodeGroupAutoscalerReconciler) Reconcile(ctx context.Context, req c
 		return RequeueWithError(r.Log, err.Error(), err)
 	}
 	current := nodeGroupAutoscaler.DeepCopy()
-	patchInstance := runtimeClient.MergeFromWithOptions(nodeGroupAutoscaler.DeepCopy(), runtimeClient.MergeFromWithOptimisticLock{})
+	patchInstance := runtimeClient.MergeFrom(nodeGroupAutoscaler.DeepCopy())
 
 	// Check if marked for deletion and run finalizers
 	if k8sutil.IsMarkedForDeletion(nodeGroupAutoscaler.ObjectMeta) {